Añadiendo confirm de delete

+++ src/pages/TasksList.jsx
@@ -15,32 +15,37 @@ function TasksList() {
     }
 
     const handleDelete = async (id) => {
-
-        try {
-            setTasks(prevTasks => prevTasks.map(task => {
-                if (task.id === id) {
-                    return { ...task, deleted: true }
-                }
-                return task;
-            }))
-            const response = await fetch(`https://tasks-node-postgres.vercel.app/tasks/${id}`, {
-                method: 'DELETE'
-            })
-
-            if (response.ok) {
-                setTasks(prevTasks => prevTasks.filter(task => task.id !== id))
-            } else {
+        const isDelete = confirm('¿Estas seguro que quieres eliminar esta tarea?')
+        if (isDelete) {
+            try {
                 setTasks(prevTasks => prevTasks.map(task => {
                     if (task.id === id) {
-                        return { ...task, deleted: false }
+                        return { ...task, deleted: true }
                     }
                     return task;
                 }))
-                console.error('Error al eliminar la tarea en el servidor');
+                const response = await fetch(`https://tasks-node-postgres.vercel.app/tasks/${id}`, {
+                    method: 'DELETE'
+                })
+
+                if (response.ok) {
+                    setTasks(prevTasks => prevTasks.filter(task => task.id !== id))
+                } else {
+                    setTasks(prevTasks => prevTasks.map(task => {
+                        if (task.id === id) {
+                            return { ...task, deleted: false }
+                        }
+                        return task;
+                    }))
+                    console.error('Error al eliminar la tarea en el servidor');
+                }
+            } catch (error) {
+                console.error(error)
             }
-        } catch (error) {
-            console.error(error)
+        } else {
+            return
         }
+
     }
 
     useEffect(() => {
@@ -50,32 +55,33 @@ function TasksList() {
     return (
         <>
             <Tasks />
-            {/* <h1>TasksList</h1> */}
-            {tasks && tasks.map(task => (
-                <article
-                    key={`${task.id}_${Math.random()}`}
-                    className='tasks__cards'>
-                    <div className='tasks'>
-                        <div className='task__title'>
-                            <p>{task.title}</p>
+            <section className='tasks__cards'>
+                {tasks && tasks.map(task => (
+                    <article
+                        key={`${task.id}_${Math.random()}`}
+                    >
+                        <div className='tasks'>
+                            <div className='task__title'>
+                                <p>{task.title}</p>
+                            </div>
+                            <div className='task__description'>
+                                <p>{task.description}</p>
+                            </div>
+                            <div className='buttons'>
+                                <button className='button'
+                                    onClick={() => navigate(`/tasks/${task.id}/edit`)}>
+                                    🔨
+                                </button>
+                                <button className='button'
+                                    onClick={() => handleDelete(task.id)}>
+                                    ❌
+                                </button>
+                            </div>
                         </div>
-                        <div className='task__description'>
-                            <p>{task.description}</p>
-                        </div>
-                        <div className='buttons'>
-                            <button className='button'
-                                onClick={() => navigate(`/tasks/${task.id}/edit`)}>
-                                🔨
-                            </button>
-                            <button className='button'
-                                onClick={() => handleDelete(task.id)}>
-                                ❌
-                            </button>
-                        </div>
-                    </div>
-                </article>
-            ))
-            }
+                    </article>
+                ))
+                }
+            </section>
 
         </>
     )
